Fusion v18.1 Update
Revised the "Gyroflow Fusion" script to use the improved Python 3 API support in Resolve/Fusion v18.1+. This removes Lua scripted "comp.Paste()" function workarounds.
